--- medtrust-back/OLD-lambda_functions/README.md
@@ -1,24 +1,0 @@
-# Lambda Functions
-
-This demonstrates how to call the AWS Lambdas function hosted for this project
-
-## API Endpoint
-
-- **Method:** `POST`
-- **URL:** `https://3hmryuz7zufnp77t7sweh4imve0lrpuq.lambda-url.us-east-1.on.aws/`
-- **Content-Type:** `application/json`
-
-## Request Example
-
-### HTTP Request
-
-```http
-POST / HTTP/1.1
-Host: 3hmryuz7zufnp77t7sweh4imve0lrpuq.lambda-url.us-east-1.on.aws
-Content-Type: application/json
-Content-Length: 27
-
-{
-    "message": "Hola"
-}
-```
