+++ medtrust-back/OLD/OLD-lambda_functions/Sonnet3-5ConverseBedrock.py
