+++ medtrust-back/OLD/OLD-vector-store-v2/demo_questions.txt
