+++ medtrust-back/OLD/OLD-vector-store-v2/faiss_index/index.faiss
