+++ medtrust-back/OLD/OLD-vector_store/faiss_index/index.pkl
