--- medtrust-back/requirements.txt
@@ -1,11 +1,0 @@
-# boto3
-# awscli
-
-pypdf 
-
-langchain 
-langchain_community
-langchain_ollama
-#langchain_aws
-
-faiss-cpu
